refactor: repository, service 패키지 변경
